implement override protection when saving program

+++ sample_service/app/Controller/ProgramsController.php
@@ -3,42 +3,47 @@ class ProgramsController extends AppController {
     public function create() {}
     public function add() {
 	if ($this->request->is('post')) {
-            $program = $this->request->data;
-	    $data_url = $this->saveProgram($program['name'], $program['serialized_data']);
-	    if ( $data_url ) {
-		$this->Program->create();
+            $data = $this->request->data['program'];
+	    $override = $this->request->data['override'] == "true";
 
-		$program['data_url'] = $data_url;
+	    $program = $this->Program->find('first', array('conditions' => array('Program.name' => $data['name'])));
 
-		if ($this->Program->save($program)) {
-		    $this->response->body('program has been saved.');
-		    return $this->response;
-		} else {
-		    $this->setSuccessFlash("fail: program save");
+	    $alreadyExists = $program != null;
+	    $response = array('success' => false, 'exists' => $alreadyExists, 'override' => $override);
+
+	    if ( !$alreadyExists || $override ) {
+		$data_url = $this->saveProgram($data['user_id'], $data['name'], $data['serialized_data'], $override);
+		if ( $data_url ) {
+		    $this->Program->create();
+
+		    $data['data_url'] = $data_url;
+		    if ( $alreadyExists && $override ) $data['id'] = $program['Program']['id'];
+		    $response['success'] = $this->Program->save($data);
 		}
 	    }
-	    else {
-		$this->setSuccessFlash("fail: program save");
-	    }
 	}
+
+	$this->response->body(json_encode($response));
+	return $this->response;
     }
 
-    private function saveProgram($name, $data) {
-	$dir = $this->getProgramDir();
+    private function saveProgram($userId, $name, $data, $override = false) {
+	$dir = $this->getProgramDir($userId);
 
 	if ( file_exists($dir) || mkdir($dir, 0777, true) ) {
 	    $path = $dir.$name;
-	    if ( file_put_contents($path, $data, LOCK_EX) ) {
-		return $path;
+	    if ( /*!file_exists($path) || $override*/true ) {
+		if ( file_put_contents($path, $data, LOCK_EX) ) {
+		    return $path;
+		}
 	    }
 	}
 
 	return false;
     }
 
-    private function getProgramDir() { 
-	$authUser = $this->getAuthUser();
-	return $authUser ? WWW_ROOT.'/files/programs/'.$authUser['id'].'/' : null; 
+    private function getProgramDir($userId) { 
+	return WWW_ROOT.'files/programs/'.$userId.'/'; 
     }
 }
 ?>
